adding in new items and bonuses. added in a temporary bonus. beginning to allow enemies to also have their own bonuses. I STILL NEED to implement bonuses that can be utilized by anyone.

+++ src/net/anorrah/MeleeWeaponItem.java
@@ -4,7 +4,7 @@ import java.util.ArrayList;
 
 public abstract class MeleeWeaponItem extends ItemObject{
 
-	double damage;
+	int damage;
 	
 	
 	public MeleeWeaponItem(int levelBonus)
